--- a/lib/tools.js
+++ b/lib/tools.js
@@ -45,7 +45,7 @@ function asString (str) {
   var found = false
   var point = 255
   const l = str.length
-  if (l > 42) {
+  if (l > 100) {
     return JSON.stringify(str)
   }
   for (var i = 0; i < l && point >= 32; i++) {
@@ -95,9 +95,26 @@ function asJson (obj, msg, num, time) {
     for (var key in obj) {
       value = obj[key]
       if ((notHasOwnProperty || obj.hasOwnProperty(key)) && value !== undefined) {
-        value = (stringifiers[key] || stringify)(serializers[key] ? serializers[key](value) : value)
-        if (value !== undefined) {
-          data += ',"' + key + '":' + value
+        value = serializers[key] ? serializers[key](value) : value
+
+        switch (typeof value) {
+          case 'undefined':
+            break
+          case 'string':
+            value = (stringifiers[key] || asString)(value)
+            if (value !== undefined) {
+              data += ',"' + key + '":' + value
+            }
+            break
+          case 'number':
+          case 'boolean':
+            data += ',"' + key + '":' + value
+            break
+          default:
+            value = (stringifiers[key] || stringify)(value)
+            if (value !== undefined) {
+              data += ',"' + key + '":' + value
+            }
         }
       }
     }