This PR does two things:

1. bring down the `deep` benchmark from 16KB per log line to 2KB. Those are still massive.
2. adds a fast path for strings, numbers and booleans so that we do not have to go in C++/`JSON.stringify`.

This is an alternative to #463 

Before:

```
benchBunyanDeepObj*10000: 2121.646ms
benchWinstonDeepObj*10000: 2946.068ms
benchBoleDeepObj*10000: 3726.740ms
benchLogLevelDeepObj*10000: 6570.217ms
benchPinoDeepObj*10000: 2721.078ms
benchPinoUnsafeDeepObj*10000: 2486.789ms
benchPinoExtremeDeepObj*10000: 2801.805ms
benchPinoUnsafeExtremeDeepObj*10000: 2590.682ms
benchBunyanDeepObj*10000: 2158.406ms
benchWinstonDeepObj*10000: 2752.963ms
benchBoleDeepObj*10000: 3433.998ms
benchLogLevelDeepObj*10000: 6406.703ms
benchPinoDeepObj*10000: 2869.529ms
benchPinoUnsafeDeepObj*10000: 2612.237ms
benchPinoExtremeDeepObj*10000: 2828.076ms
benchPinoUnsafeExtremeDeepObj*10000: 2547.889ms
```

After:

```
benchBunyanDeepObj*10000: 2041.287ms
benchWinstonDeepObj*10000: 2883.688ms
benchBoleDeepObj*10000: 3711.500ms
benchLogLevelDeepObj*10000: 6490.075ms
benchPinoDeepObj*10000: 2248.998ms
benchPinoUnsafeDeepObj*10000: 2223.362ms
benchPinoExtremeDeepObj*10000: 2316.757ms
benchPinoUnsafeExtremeDeepObj*10000: 2304.686ms
benchBunyanDeepObj*10000: 2165.046ms
benchWinstonDeepObj*10000: 2776.695ms
benchBoleDeepObj*10000: 3393.842ms
benchLogLevelDeepObj*10000: 6362.907ms
benchPinoDeepObj*10000: 2265.468ms
benchPinoUnsafeDeepObj*10000: 2267.382ms
benchPinoExtremeDeepObj*10000: 2272.065ms
benchPinoUnsafeExtremeDeepObj*10000: 2295.960ms
```

However this also improves the basic `'object'` benchmark:

before

```
benchBunyanObj*10000: 669.430ms
benchWinstonObj*10000: 562.287ms
benchBoleObj*10000: 574.375ms
benchLogLevelObject*10000: 558.069ms
benchPinoObj*10000: 299.204ms
benchPinoUnsafeObj*10000: 302.748ms
benchPinoExtremeObj*10000: 177.923ms
benchPinoUnsafeExtremeObj*10000: 161.237ms
benchBunyanObj*10000: 611.520ms
benchWinstonObj*10000: 474.447ms
benchBoleObj*10000: 325.590ms
benchLogLevelObject*10000: 520.530ms
benchPinoObj*10000: 281.796ms
benchPinoUnsafeObj*10000: 276.001ms
benchPinoExtremeObj*10000: 146.786ms
benchPinoUnsafeExtremeObj*10000: 146.729ms
```

after

```
benchBunyanObj*10000: 675.986ms
benchWinstonObj*10000: 565.977ms
benchBoleObj*10000: 594.813ms
benchLogLevelObject*10000: 583.699ms
benchPinoObj*10000: 277.615ms
benchPinoUnsafeObj*10000: 277.089ms
benchPinoExtremeObj*10000: 139.002ms
benchPinoUnsafeExtremeObj*10000: 130.699ms
benchBunyanObj*10000: 674.169ms
benchWinstonObj*10000: 490.112ms
benchBoleObj*10000: 323.424ms
benchLogLevelObject*10000: 520.541ms
benchPinoObj*10000: 244.344ms
benchPinoUnsafeObj*10000: 246.856ms
benchPinoExtremeObj*10000: 123.631ms
benchPinoUnsafeExtremeObj*10000: 122.023ms
```