@jsumners 2018-08-02
This looks far simpler. I like it.